requirements.txt
matplotlib==3.10.0
phi==0.6.7
phidata==2.7.6
python-dotenv==1.0.1
Requests==2.32.3
supabase==2.11.0
